+++ src/main.cpp
@@ -11,6 +11,7 @@
 #include "NavigatorEngine.hpp"
 #include "OdometryEngine.hpp"
 #include "UserInputDriver.hpp"
+#include "NavigatorGraphBuilder.hpp"
 
 using namespace std;
 using namespace Manhattan;
@@ -31,8 +32,9 @@ int main(const int argc, char* argv[])
     app->AddComponent<Core::UserInputDriver>();
     app->AddComponent<Core::MappingEngine>();
     app->AddComponent<Core::NavigatorEngine>();
-    app->AddComponent<Core::FollowerEngine>()->Enable();
-    app->AddComponent<Core::ExplorerEngine>();//->Enable();
+    app->AddComponent<Core::FollowerEngine>();
+    app->AddComponent<Core::ExplorerEngine>()->Enable();
+    app->AddComponent<Core::NavigatorGraphBuilder>();
 
     app->Run();